Add documentation and implement RandomAccessListInterface for efficient algorithms
- Introduced README.md with architecture notes and test instructions.
- Created ComparatorResolver for consistent comparator handling.
- Enhanced AbstractDataset to validate CSV header alignment.
- Updated Dijkstra's algorithm to utilize a factory for priority queues.
- Refactored Graph class to use Edge objects and improved vertex management.
- Added RandomAccessAdapter to ensure compatibility with random-access algorithms.
- Implemented RandomAccessListInterface and updated ArrayList accordingly.
- Modified PriorityQueue to use a binary heap for improved performance.
- Updated search and sort algorithms to utilize RandomAccessAdapter for better efficiency.

+++ src/Lists/Interfaces/PriorityQueueInterface.php
@@ -6,7 +6,7 @@ namespace Algoritmes\Lists\Interfaces;
 
 interface PriorityQueueInterface
 {
-    public function enqueue(mixed $item, int $priority): void;
+    public function enqueue(mixed $item, float $priority): void;
     public function dequeue(): mixed;
     public function isEmpty(): bool;
     public function size(): int;
